Add response cookie module and update HttpResponse initialization
Introduce a new `response_cookie` module containing the `Cookie` and `CookieOptions` structs for managing HTTP cookies. Update the `HttpResponse` initialization in tests to use `HttpResponse::new()` instead of the deprecated `HttpResponse::default()`, ensuring consistency across the test suite.
